backend/mysite/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
Django==3.0.3
django-filter==2.3.0
django-mail-auth==2.1.0
django-modelcluster==5.0.2
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1596632844333/work
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446511143/work
jdcal==1.4.1
l18n==2018.5
Markdown @ file:///C:/ci/markdown_1597415206643/work
openpyxl==3.0.4
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
Unidecode==1.1.1
urllib3==1.25.10
wagtail==2.10
webencodings==0.5.1
Willow==1.4
wincertstore==0.2
xlrd==1.2.0
XlsxWriter==1.3.3
xlwt==1.3.0
zipp==3.1.0
